[CompositeClipPathAnimation] Resolve flakey tests.
Mac failures of the fixed-position test have stopped and the test can now be re-enabled. Two-clip-path-animation-diff-length1 was flakey as sometimes animations[1] would not have populated by the time the scripts were run. This change adds a RAF that resolves the issue, and it no longer repros with a 1000 executions (previously the repro rate was approximately 1:100)

Bug: 1474966, 1411760

Change-Id: I6bf988254d9bf8cc042dc54cff63598513627e61