Add WPT test validating that drawImage writes to the layer
Added a WPT to check that using `drawImage` inside a layer writes the
image to the layer and not it's parent directly. The test uses
`drawImage` with another canvas a source, inside a layer that applies a
filter. In its reference, the same drawing commands are used in the main
canvas inside the layer, to verify that the filter effect is applied on
the image.

Bug: 1396372
Change-Id: I377b4eb0b77865c639b296df70110c2fa8f7f07f