Clear user activation on cross-site navigations.
When full site isolation is disabled, renderer processes and frame
instances are re-used when performing cross-site navigations. This
includes user activation state, and, more specifically, the sticky
`has_been_active_` bit in `UserActivationState`.

Currently, the `UserActivationState` on the renderer-side is reset only
if the navigation's associated frame is a main frame. That means that if
an iframe navigates to a cross-site page, its sticky user activation
state will be the leftover state from the previous page. So, if a user
interacted with the previous page in any capacity, the newly loaded page
will think it has received a user gesture, essentially using an
unintentional cache of the user activation state.

This becomes an issue when dealing with our framebusting interventions.
We only allow an iframe to do a top-level navigation if it received a
user gesture. However, if an iframe's previous page load received a user
activation, or worse, if the iframe was not navigated to anything and
got a user activation because its embedder was interacted with, this
allows the current page to circumvent our framebusting interventions.
The latter happens because of same-origin descendant activation
behavior. See:
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/frame_tree_node.cc;l=766-778;drc=30753b1135fa271a3b45bbdbfef6567e46733a7f;bpv=1;bpt=1

Note that this problem does not exist if site isolation is enabled (site
isolation is enabled by default on desktop platforms), since a
cross-site navigation will create a whole new process with a fresh
`UserActivationState`.

To fix this, this CL clears the user activation state on cross-site
subframe navigations in the renderer (user activation is already cleared
for main frames). This CL also cleans up the browser-side user
activation clearing code to match what the renderer is doing.

The ultimate end goal is to unconditionally clear the user activation
state for all navigations. That unfortunately is not possible today as
there are entrenched use cases that rely on sticky user activation state
being cached for same-site navigations. See: https://crbug.com/40228985.

Bug: 41493458
Change-Id: Ibec11437fcd03470571e04a4e0dfaadffddf6c03