[shared storage] Introduce WorkletOptions param
Introduce WorkletOptions param for both addModule() and
createWorklet(). For the resource request, use the provided
credentials and use the SiteForCookies derived from the RFHI. This
aligns with the worklet standard [1], and aligns with the
implementation for other workers (e.g. for dedicated workers [2]).

Note that the new param for addModule() isn't gated by
kSharedStorageAPIM123. This is intentional: we have been deviating
from the worklet standard, so we consider this change as a bugfix.

[1] https://html.spec.whatwg.org/multipage/worklets.html#dom-worklet-addmodule
[2] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/worker_host/dedicated_worker_host.cc?q=WorkerScriptFetcher::CreateAndStart

Bug: 1218540
Change-Id: I82bb7f7509e904f1d9e2ff26c9c8eb1b90ad5828