[Event Timing] Exposed interactionId to Keypress & keydown/up under composition
Changed the logic of generating interaction Id for keyboard events.
The interaction id is now generated on the keydown entry.
As the keypress event always follows the keydown event,
the same (CurrentInteractionId) is assigned to the keypress event. For composition events interaction id is updated on each input/keydown when compositionupdate appeared.

They kCompositionend always has .isTrusted=false that impacts the transition between composition and noncomposition events. I updated the event_timing logic, however the best solution would be to fix the bug:
https://bugs.chromium.org/p/chromium/issues/detail?id=1038569

Bug: 1456384
Change-Id: Ie736f3f189e9a349badbe03c4a65627ac491eff4