Add 'writingSuggestions' attribute
This CL implements the new 'writingSuggestions' attribute. We add a new
runtime flag and setter/getter methods for it.

The attribute has values "true"/"false" to allow developers to enable or
disable browser-provided writing suggestions. If the attribute is not
explicitly set on an element, it inherits its value from ancestor
elements if specified; otherwise, it defaults to "true".

Changes were verified against wpt tests:
(note: checks for getAttribute() for missing/invalid attributes differ
to align with default behavior)
https://github.com/web-platform-tests/wpt/pull/43780

Intent to Prototype:
https://groups.google.com/a/chromium.org/g/blink-dev/c/rHyRCx-hJhE

Bug: 1513953
Change-Id: I716deba02f3d8df37dd04d1363d624cf2615203c