Convert existing orientation-event web tests to test_driver
The purpose of this change is to stop using the JS mocks-based
implementation and start using WebDriver and virtual sensors for
interoperability. WebDriver commands for Generic Sensor are defined in
https://www.w3.org/TR/generic-sensor/#automation, and the Device
Orientation-specific parts are defined in
https://www.w3.org/TR/orientation-event/#automation.

The TestHelper class was created to simplify tests. It takes care of
creating and removing the backing virtual sensors for each event type,
and handling permissions. It also contains some previously free
functions that now depend on virtual sensors (such as setData()).
More generic functions which do not need information saved in TestHelper
remain untouched in the same orientation-event-helpers.js file.

Care has been taken to avoid changing the values and the order of the
calls in the tests as much as possible. Notes about some cases where it
was not possible:
1. Currently it is not possible to set individual values to null because
     the parsing algorithms used by
     https://w3c.github.io/sensors/#update-virtual-sensor-reading-command
     always expect numbers. This affected null-values.https.html test.
2. "deviceorientation" and "deviceorientationabsolute" events whose
    readings don't differ from the previous one by a certain amount are
    ignored per spec (https://www.w3.org/TR/orientation-event/#significant-change-in-orientation).
     Because of this requirement second test values needed to be changed
     in add-listener-from-callback.https.html test for Device Orientation.
3. motion/add-listener-from-callback.https.html: The second listener is
   added before the first one is removed to avoid needlessly stopping
   the sensor.

Bug: 1520919
Change-Id: I7890f08357481fd62a71338b20950b534abbbe8d