taskcluster: update wpt docker image to use Ubuntu 22.04 as base
The wpt taskcluster docker image currently uses ubuntu 20.04 as the
base. This means any binary that we run using this image must be
compatible with the glibc version in Ubuntu 20.04.

As reported in #43361, this blocks Servo from upgrading our own CI
runners (Github Actions based) to the latest Ubuntu images, since the
browser binary produced will link against newer GlibC. We had worked
around this in Servo by using 20.04 runners only for the job that built
servo binary used by WPT.fyi.

Due to actions/runner-images#11101, it is now urgent for Servo to
move away from 20.04.

This patch simply attempts to switch the base of the taskcluster image
to Ubuntu 22.04. Some prelimarily santiy checks were done locally:
 - The new image builds successfully
 - A subset of Servo reftests were run using the new local image
   and validated to contain the same number of PASSES

I have not attempted to move to 24.04 directly as there were issues
raised in #43361 related to wayland-as-default in 24.04. I don't have
any insight into such issues, so I'm updating the image to only 22.04
for now, as a quick fix.

**Before merging this patch, more thorough testing is required**. This
requires:

 - [ ] Publishing the updated docker image to Github Container registry
 - [ ] Triggering a test run in the staging version of WPT.fyi with the
       updated docker image and ensuring the image works with all
       supported browsers.

I don't have the permissions needed to publish the image and also lack
have full knowledge of the WPT infra to perform such testing. I hope
someone with such acess can help me with this.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>